Major bug fix for mask+NCC metric; initial code for multi chunk greedy.
Fixed a major issue that was introduced with the new WNCC metric code.
When WNCC was added, the original NCC behavior was changed so that the
fixed image mask was used both as the mask for integrating the
pixel-wise metric and to mask the fixed image for the computation of the
metric. As the result, when the mask came up to the edge of the tissue
in the image, the metric would have high values at the borders and
registration would behave poorly. The original behavior, where the mask
was dilated so that the fixed image is sampled at a wider band than the
metric integration mask is now restored. Additionally, the option to
dilate the fixed/moving masks for WNCC has been finally implemented.

Added the first draft of multi chunk greedy registration, which uses a
multi-label mask as input and performs parallel rigid registrations for
each chunk. The intention is to add code for constraining the rigid
registrations to penalize overlap between chunks.