Fix WNCC major bug with mask premultiplication
The WNCC metric is coded under the assumption that the inputs are
intensity*mask and mask, but this multiplication of intensity by
mask wask not happening anywhere in the code. This seems to be
messing with WNCC gradient near the edges of the moving image so
the registration was not quite happening there. Not sure how long
this has been going on for, might have been since the introduction
of the WNCC metric. Noted a significant improvemnt near mask edges
after making this change.